+++ src/frontend/js/components/SaleTunnel/SaleTunnelSuccess/index.tsx
@@ -10,33 +10,33 @@ const messages = defineMessages({
   congratulations: {
     defaultMessage: 'Congratulations!',
     description: 'Text displayed to thank user for his order',
-    id: 'components.SaleTunnelStepResume.congratulations',
+    id: 'components.SaleTunnelSuccess.congratulations',
   },
   successMessage: {
     defaultMessage: 'Your order has been successfully created.',
     description: 'Message to confirm that order has been created',
-    id: 'components.SaleTunnelStepResume.successMessage',
+    id: 'components.SaleTunnelSuccess.successMessage',
   },
   successDetailMessage: {
     defaultMessage: 'You will receive your invoice by email in a few moments.',
     description: "Text to remind that order's invoice will be send by email soon",
-    id: 'components.SaleTunnelStepResume.successDetailMessage',
+    id: 'components.SaleTunnelSuccess.successDetailMessage',
   },
   successDetailSignatureMessage: {
     defaultMessage:
       'In order to enroll to course runs you first need to sign the training contract.',
     description: 'Text to remind that order needs to be signed',
-    id: 'components.SaleTunnelStepResume.successDetailSignatureMessage',
+    id: 'components.SaleTunnelSuccess.successDetailSignatureMessage',
   },
   cta: {
     defaultMessage: 'Start this course now!',
     description: 'Label to the call to action to close sale tunnel',
-    id: 'components.SaleTunnelStepResume.cta',
+    id: 'components.SaleTunnelSuccess.cta',
   },
   ctaSignature: {
     defaultMessage: 'Sign the training contract',
     description: 'Label to the call to action to close sale tunnel if there is a pending signature',
-    id: 'components.SaleTunnelStepResume.ctaSignature',
+    id: 'components.SaleTunnelSuccess.ctaSignature',
   },
 });
 
@@ -44,14 +44,14 @@ export const SaleTunnelSuccess = ({ closeModal }: { closeModal: () => void }) =>
   const intl = useIntl();
   const { order, product } = useSaleTunnelContext();
   return (
-    <section className="SaleTunnelSuccess" data-testid="generic-sale-tunnel-success-step">
-      <header className="SaleTunnelSuccess__header">
+    <section className="sale-tunnel-end" data-testid="generic-sale-tunnel-success-step">
+      <header className="sale-tunnel-end__header">
         <SuccessIcon />
-        <h3 className="SaleTunnelSuccess__title">
+        <h3 className="sale-tunnel-end__title">
           <FormattedMessage {...messages.congratulations} />
         </h3>
       </header>
-      <p className="SaleTunnelSuccess__content">
+      <p className="sale-tunnel-end__content">
         <FormattedMessage {...messages.successMessage} />
         <br />
         <FormattedMessage {...messages.successDetailMessage} />
@@ -62,7 +62,7 @@ export const SaleTunnelSuccess = ({ closeModal }: { closeModal: () => void }) =>
           </>
         )}
       </p>
-      <footer className="SaleTunnelSuccess__footer">
+      <footer className="sale-tunnel-end__footer">
         {product.contract_definition ? (
           <Button
             href={