+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Add `SaleTunnelNotValidated` step into `SaleTunnel`
 - Add WarningIcon component
 
 ### Changed